# openapi.yaml
openapi: 3.0.0
info:
  title: '77sol case'
  version: '0.1'
paths:
...
  /cliente/cadastrar:
    get:
      tags:
        - Clientes
      summary: 'Exibir o form de cadastro de clientes'
      operationId: 627f750c358e74ddb541b439aa73b8bb
      responses:
        '200':
          description: 'Exibir o form de cadastro de clientes'
        '401':
          description: 'Não autorizado'
  '/cliente/editar/{id}':
    get:
      tags:
        - Clientes
      summary: 'Exibir o form de editar o cliente'
      operationId: 1ad44f7b9dbc54fde132aeb0ef2ede00
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Exibir o form de editar o cliente'
        '401':
          description: 'Não autorizado'
  /cliente/salvar:
    post:
      tags:
        - Clientes
      summary: 'Cadastrar um novo cliente'
      operationId: 7ff7ad04e6c261385fcd4b8219b1c28b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - nome
                - email
                - telefone
                - cpf_cnpj
              properties:
                nome:
                  type: string
                  example: 'Pedro da Silva'
                email:
                  type: string
                  example: pedro.silva@example.com
                telefone:
                  type: string
                  example: '11944448888'
                cpf_cnpj:
                  type: string
                  example: '12345678911'
              type: object
      responses:
        '201':
          description: 'Cliente cadastrado com sucesso'
        '400':
          description: 'Requisição inválida'
  '/api/cliente/delete/{id}':
    delete:
      tags:
        - Clientes
      summary: 'Deletar um cliente'
      operationId: 3c9c2f32fea732a4ed99b4f5be59ef31
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Cliente deletado com sucesso'
        '500':
          description: 'Ocorreu um erro ao excluir o cliente'
...